add filtering, ordering and pagination, todo: convert tests to snake_case

+++ run_tests.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+shopt -s nullglob
+
 TEST_DIRS=(
     tests/graphql/actions
     tests/graphql/django_objects
@@ -8,6 +10,14 @@ TEST_DIRS=(
 )
 
 cd "${0%/*}"
+
+# testcases filtering by regex
+if [[ -n "$1" ]]
+then
+    filter=$1
+else
+    filter="*"
+fi
 
 # activate virtualenv if exists
 if [[ -d venv ]]
@@ -19,7 +29,13 @@ for dir in "${TEST_DIRS[@]}"
 do
     for tc in "$dir"/[!_]*
     do
+        # ignore if the testcase does not match the filter
+        if ! grep -E "$filter" <<< "$tc" &>/dev/null
+        then
+            continue
+        fi
         echo "Running $dir/$tc"
+
         # cleanup possible old runs
         find test_project/simple_api/testcases ! \( -name '__init__.py' -o -name 'admin.py' -o -name 'apps.py' -o -name 'utils.py' -o -name 'views.py' \) -type f -exec rm -rf {} \;
         rm -f test_project/simple_api/db.sqlite3
